# Improve json encoding performance for specs-go