… specs-go/config_ffjson.go
… specs-go/state_ffjson.go