@vbatts 2016-03-16
Then ought we put a target in the `Makefile` or go generate in the other source?