@marcosnils 2016-03-25
@crosbymichael @vbatts time to move to easyjson?. https://github.com/mailru/easyjson